backend/requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
pillow>=8.3.1
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.21.1
python-dotenv>=0.19.0
werkzeug>=2.0.1
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
gunicorn>=20.1.0
google-generativeai>=0.3.0
requests>=2.31.0